regex: actually checking the proof, fixes

+++ src/include/gnunet_regex_lib.h
@@ -119,7 +119,7 @@ GNUNET_REGEX_eval (struct GNUNET_REGEX_Automaton *a,
 
 /**
  * Get the first key for the given 'input_string'. This hashes
- * the first x bits of the 'input_strings'.
+ * the first x bits of the 'input_string'.
  *
  * @param input_string string.
  * @param string_len length of the 'input_string'.
@@ -128,18 +128,18 @@ GNUNET_REGEX_eval (struct GNUNET_REGEX_Automaton *a,
  * @return number of bits of 'input_string' that have been consumed
  *         to construct the key
  */
-unsigned int /* FIXME: size_t */
-GNUNET_REGEX_get_first_key (const char *input_string, /* FIXME: size_t */ unsigned int string_len,
+size_t
+GNUNET_REGEX_get_first_key (const char *input_string, size_t string_len,
                             struct GNUNET_HashCode * key);
 
 
 /**
  * Check if the given 'proof' matches the given 'key'.
  *
- * @param proof partial regex
- * @param key hash
+ * @param proof partial regex of a state.
+ * @param key hash of a state.
  *
- * @return GNUNET_OK if the proof is valid for the given key
+ * @return GNUNET_OK if the proof is valid for the given key.
  */
 int
 GNUNET_REGEX_check_proof (const char *proof,
